move aggregation related files to filestudy utils
