+++ antarest/study/storage/rawstudy/model/filesystem/matrix/output_series_matrix.py
@@ -20,7 +20,7 @@ from typing_extensions import override
 from antarest.core.exceptions import ChildNotFoundError, MustNotModifyOutputException
 from antarest.core.model import JSON
 from antarest.matrixstore.matrix_uri_mapper import add_matrix_id_prefix
-from antarest.output.utils import get_start_column, parse_output_file
+from antarest.output.filestudy.utils import get_start_column, parse_output_file
 from antarest.study.model import MatrixFrequency
 from antarest.study.storage.rawstudy.model.filesystem.config.model import FileStudyTreeConfig
 from antarest.study.storage.rawstudy.model.filesystem.lazy_node import LazyNode
